+++ routers/admin/__init__.py
@@ -2,7 +2,7 @@ from aiogram import Router
 
 from .admin_commands import router as admin_commands_router
 from .actions_handler import router as actions_handler_router
-from .phone_handler import router as phone_handler_router
+from .contact_handler import router as phone_handler_router
 
 
 router = Router(name='admin')
